Big refactor to keep a similar schema, but separate generic PING from my research.
