+++ snps.py
@@ -25,15 +25,15 @@ if sys.argv[1] in ['view', 'download']:
         print("Loading SNPS...")
         snp_metadata = sess.get_snp_metadata(snp)
         print("Found %d SNPs for %s" % (snp_metadata is not None, snp))
-        print(snp_metadata)
+        print('\n'.join([str(snp) for snp in snp_metadata]))
 
         print("Loading genes...")
         all_genes = sess.get_genes_from_snp(snp)
         print("Found %d genes for %s" % (all_genes is not None, snp))
-        print(all_genes)
+        print('\n'.join([str(gene) for gene in all_genes]))
 
         print("GWAS results:")
-        print('\n'.join(sess.snp_gwas_results(snp)))
+        print('\n'.join([str(gwas) for gwas in sess.snp_gwas_results(snp)]))
 
     else:
         gene_name = sys.argv[2]
@@ -41,11 +41,12 @@ if sys.argv[1] in ['view', 'download']:
         print("Loading genes...")
         gene_metadata = sess.get_gene_metadata(gene_name)
         print("Found %d genes for %s" % (len(gene_metadata), gene_name))
-        print(gene_metadata)
+        print('\n'.join([str(gene) for gene in gene_metadata]))
 
         print("Loading snps...")
         all_snps = sess.get_snps_from_gene(gene_metadata)
         print("Found %d snps for %s" % (len(all_snps), gene_name))
+        print('\n'.join([str(snp) for snp in all_snps]))
 
         if sys.argv[1] == 'download':
             sess.login()
